+++ pkg/processor/procjs/builder.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package jsprocessor
+package procjs
 
 import (
 	"github.com/conduitio/conduit/pkg/foundation/cerrors"
